pull before

+++ QQEgg_Backend/DTO/PsiteRoomDTO.cs
@@ -7,8 +7,8 @@ namespace QQEgg_Backend.DTO
         public int RoomId { get; set; }        
         public int? SiteId { get; set; }
         public int? CategoryId { get; set; }
-        public decimal? HourPrice { get; set; }
-        public decimal? DatePrice { get; set; }
+        public int? HourPrice { get; set; }
+        public int? DatePrice { get; set; }
         public int? Ping { get; set; }
         public string? Image { get; set; }
         public bool? Status { get; set; }
