+++ QQEgg_Backend/Models/TAorders.cs
@@ -13,7 +13,7 @@ namespace QQEgg_Backend.Models
         public DateTime OrderDate { get; set; }
         public DateTime? EndDate { get; set; }
         public int? Clicks { get; set; }
-        public decimal? Price { get; set; }
+        public int? Price { get; set; }
 
         public virtual TAdvertise Advertise { get; set; }
         public virtual TSuppliers Supplier { get; set; }
